+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New dataloader for classification metrics (precision, recall, f1, support)
 - Standard deviation for run latency in experiments table
 - Sorting feature in experiments table
+- Resizing columns in experiments table
 
 ### Changed
 - Updated the `ProjectPageHeader` component to display classification metrics